Created a client callback, tweaking the Message queue with the new remote objects and methods
