+++ src/main/java/Client/Client.java
@@ -1,0 +1,7 @@
+package Client;
+
+public class Client {
+	public static void main(String[] args) {
+		ClientCallback callback = new ClientCallbackImpl();
+	}
+}
