+++ src/main/java/Client/ClientCallback.java
@@ -1,0 +1,11 @@
+package Client;
+
+import java.rmi.RemoteException;
+
+import util.Attacks.BasicWebAttack;
+
+public interface ClientCallback extends java.rmi.Remote {
+	
+	void update(BasicWebAttack attacks) throws RemoteException;
+
+}
