+++ src/main/java/messagequeue/RMIMessageQueue.java
@@ -4,9 +4,11 @@ import util.Attacks;
 
 import java.rmi.RemoteException;
 
+import Client.ClientCallback;
+
 public interface RMIMessageQueue extends java.rmi.Remote {
 
-	void createTask(Attacks.BasicWebAttack attack) throws java.rmi.RemoteException;
+	void createTask(Attacks.BasicWebAttack attack, ClientCallback clientCallback) throws java.rmi.RemoteException;
 
     Attacks.BasicWebAttack getTask() throws java.rmi.RemoteException;
